requirements.txt
flask
google-generativeai
Flask-Cors